requirements.txt
pytorch-lightning
sentence_transformers
torch~=1.11.0
wget~=3.2
nltk~=3.7
numpy~=1.22.0
pandas~=1.4.2
faiss-cpu
tqdm~=4.64.0
transformers~=4.18.0